+++ packages/testbot/bot.js
@@ -1,10 +1,8 @@
 const { Botkit } = require('botkit');
-const { ShowTypingMiddleware } = require('botbuilder');
 const { MongoDbStorage } = require('botbuilder-storage-mongodb');
-const basicAuth = require('express-basic-auth');
 const { BotkitCMSHelper } = require('botkit-plugin-cms');
 
-// const { SlackAdapter, SlackMessageTypeMiddleware, SlackIdentifyBotsMiddleware, SlackEventMiddleware } = require('botbuilder-adapter-slack');
+const { SlackAdapter, SlackMessageTypeMiddleware, SlackIdentifyBotsMiddleware, SlackEventMiddleware } = require('botbuilder-adapter-slack');
 // const { WebexAdapter } = require('botbuilder-adapter-webex');
 const { WebAdapter } = require('botbuilder-adapter-web');
 // const { FacebookAdapter, FacebookEventTypeMiddleware } = require('botbuilder-adapter-facebook');
@@ -50,10 +48,10 @@ if (process.env.MONGO_URI) {
  * ----------------------------------------------------------------------
 //  */
 // const adapter = new SlackAdapter({
-//     enable_incomplete: true,
-//     // verificationToken: process.env.verificationToken,
+//     // enable_incomplete: true,
+//     verificationToken: process.env.verificationToken,
 //     clientSigningSecret: process.env.clientSigningSecret,  
-//     // botToken: process.env.botToken,
+//     botToken: process.env.botToken,
 //     // clientId: process.env.clientId,
 //     // clientSecret: process.env.clientSecret,
 //     // scopes: ['bot'],
@@ -103,7 +101,7 @@ if (process.env.MONGO_URI) {
 
 // // Use SlackMessageType middleware to further classify messages as direct_message, direct_mention, or mention
 // adapter.use(new SlackMessageTypeMiddleware());
-// 
+
 /* ----------------------------------------------------------------------
  *  __      __      ___.                        __           __   
  * /  \    /  \ ____\_ |__   __________   ____ |  | __ _____/  |_ 
@@ -145,7 +143,8 @@ const adapter = new WebAdapter({});
 const controller = new Botkit({
     debug: true,
     webhook_uri: '/api/messages',
-    disable_console: true,
+    webserver_middlewares: [(req, res, next) => { console.log('REQ > ', req.url); next(); }],
+    // disable_console: true,
     adapter: adapter,
     // disable_webserver: true,
     // adapterConfig: {
@@ -156,7 +155,7 @@ const controller = new Botkit({
 });
 
 const cms = new BotkitCMSHelper({
-    uri: process.env.cms_uri,
+    cms_uri: process.env.cms_uri,
     token: process.env.cms_token,
 });
 
@@ -187,6 +186,11 @@ controller.ready(() => {
     // controller.usePlugin(require('./plugins/verbose/index.js'));
 
     if (controller.webserver) {
+
+        controller.webserver.get('/', (req, res) => {
+            res.send(controller.version);
+        });
+
 
         controller.webserver.get('/install', (req, res) => {
             // getInstallLink points to slack's oauth endpoint and includes clientId and scopes