+++ packages/testbot/features/convo.js
@@ -7,19 +7,17 @@ module.exports = function(controller) {
 
     welcome.say('Hey!');
     welcome.ask({
-        text: ['Check this out...'],
-        channelData: { 
-            quick_replies: [
-                {
-                    title: 'Foo',
-                    payload: 'foo',
-                },
-                {
-                    title: 'Bar',
-                    payload: 'bar',
-                }
-            ],
-        },
+        text: ['Check this out...', 'YO! Check it.', 'HELLO HUME...'],
+        quick_replies: [
+            {
+                title: 'Foo',
+                payload: 'foo',
+            },
+            {
+                title: 'Bar',
+                payload: 'bar',
+            }
+        ],
     },async(answer, convo, bot) => {
         // noop.
     },{key: 'waterfall_results'});
@@ -31,6 +29,11 @@ module.exports = function(controller) {
         // as does convo.vars.name
     },{key: 'name'});
 
+    welcome.ask('Tell me something! I\'ll hear it but won\'t save the value.', async (answer, convo, bot) => {
+        // noop
+        // answer contains the user's answer
+        // answer can be used but is not persisted into the convo.vars
+    });
 
     welcome.say('hrrm!');
     welcome.say('ok, {{vars.name}}!');
@@ -66,8 +69,9 @@ module.exports = function(controller) {
     welcome.addMessage('NOOOOOO', 'bar');
 
     welcome.onChange('name', async(response, convo, bot) => {
-        await bot.say('NO! NEVER!!!');
-        await convo.gotoThread('bar');
+        // await bot.say('NO! NEVER!!!');
+        // await convo.gotoThread('bar')
+        await bot.say('You changed your name.');
     });
 
     welcome.before('foo', async(convo, bot) => {